feat: restructure the whole halo2 pkgs file structure
