+++ Cargo.toml
@@ -4,10 +4,10 @@ default-members = ["pkgs/cli"]
 members = [
     "pkgs/cli",
     "pkgs/merkle-tree-wasm",
-    "pkgs/binary-merkle-tree-halo2",
     "pkgs/spartan-ecdsa-wasm", 
-    "pkgs/circuits/halo2wrong-ecdsa", 
-    "pkgs/circuits/halo2-secp256k1-wasm"
+    "pkgs/halo2wrong-ecdsa", 
+    "pkgs/halo2-ecdsa",
+    "pkgs/halo2-binary-merkle-tree"
 ]
 
 [profile.release]
