--- pkgs/binary-merkle-tree-halo2/README.md
@@ -1,13 +1,0 @@
-# @anonklub/binary-merkle-tree-halo2
-
-## Overview
-
-`@anonklub/binary-merkle-tree-halo2` is a A rust crate implements binary merkle tree using halo2 Secp256k1 F, and PSE poseidon. Designed specifically for handling anonymity sets. This crate is using `wasm-pack` crate in order to generate wasm TS liberary in order to be integrated in the client-side.
-
-## Build
-
-To build the project from source, navigate to the project directory and run the following command:
-
-```shell
-pnpm run compile
-```
