+++ pkgs/halo2-binary-merkle-tree/README.md
@@ -1,0 +1,13 @@
+# @anonklub/halo2-binary-merkle-tree
+
+## Overview
+
+`@anonklub/halo2-binary-merkle-tree` is a A rust crate implements binary merkle tree using halo2 Secp256k1 F, and PSE poseidon. Designed specifically for handling anonymity sets. This crate is using `wasm-pack` crate in order to generate wasm TS liberary in order to be integrated in the client-side.
+
+## Build
+
+To build the project from source, navigate to the project directory and run the following command:
+
+```shell
+pnpm run compile
+```
