+++ pkgs/halo2-binary-merkle-tree/package.json
@@ -1,11 +1,11 @@
 {
-    "name": "@anonklub/halo2-ecdsa-wasm",
+    "name": "@anonklub/halo2-binary-merkle-tree",
     "version": "1.1.3",
     "author": "0xisk <0xisk@proton.me>",
-    "description": "Wasm Rust crate for verifying Secp256k1 signatures",
+    "description": "A rust crate implements binary merkle tree using halo2 Secp256k1 F, and PSE poseidon.",
     "keywords": [
       "wasm",
-      "ecdsa",
+      "merkle-tree",
       "Secp256k1",
       "rust"
     ],
