+++ pkgs/halo2-binary-merkle-tree/src/binary_merkle_tree.rs
