+++ pkgs/halo2-binary-merkle-tree/src/gadget.rs
