+++ pkgs/halo2-ecdsa/package.json
@@ -1,11 +1,11 @@
 {
-    "name": "@anonklub/binary-merkle-tree-halo2",
+    "name": "@anonklub/halo2-ecdsa",
     "version": "1.1.3",
     "author": "0xisk <0xisk@proton.me>",
-    "description": "A rust crate implements binary merkle tree using halo2 Secp256k1 F, and PSE poseidon.",
+    "description": "Wasm Rust crate for verifying Secp256k1 signatures",
     "keywords": [
       "wasm",
-      "merkle-tree",
+      "ecdsa",
       "Secp256k1",
       "rust"
     ],
