+++ pkgs/halo2-ecdsa/src/circuits/mod.rs
@@ -1,0 +1,2 @@
+pub mod efficient_ecdsa;
+pub mod ecdsa;
